requirements.txt
alembic==1.8.0
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==1.1.0
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
SQLAlchemy==1.4.37
temp-email-validator==0.1
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
